Add files via upload
The zipped csv file included in this repo contains information about loans and bank decisions on those loans. This python file contains several classes which perform a diversity of functions. It includes a class with a method to iterate through a large zipped csv file, with a parameter that allows for rows to be skipped. The Loan class creates loan objects with convenient representations for their specific attributes. Bank ties these two previous classes together. DTree creates a decision tree to determine the approval of the loans. Finally, there is a randomforest class and a bias testing class. The notebook contains basic instantiations of all the classes.